refactor(navigation): created separate packages for RTDB and CF
