+++ packages/content-rtdb/src/index.ts
@@ -270,8 +270,8 @@ export const factory: FlamelinkFactory = context => {
   return api
 }
 
-const register: App.SetupModule = (context: App.Context) => {
-  if (context.dbType === 'cf') {
+export const register: App.SetupModule = (context: App.Context) => {
+  if (context.dbType === 'rtdb') {
     return factory(context)
   }
