+++ packages/flamelink/cf/navigation/index.ts
@@ -1,0 +1,1 @@
+import '@flamelink/sdk-navigation-cf'
