+++ packages/flamelink/cf/navigation/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "flamelink/cf/navigation",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "typings": "../../public.d.ts"
+}
