+++ packages/flamelink/navigation/index.cdn.ts
@@ -8,18 +8,5 @@
  * the module package.
  */
 
-import flamelink from '@flamelink/sdk-app'
-import * as App from '@flamelink/sdk-app-types'
-
-import getApiForRTDB from '@flamelink/sdk-navigation/dist/rtdb'
-import getApiForCF from '@flamelink/sdk-navigation/dist/cf'
-
-const navigation: App.SetupModule = (context: App.Context) => {
-  if (context.dbType === 'rtdb') {
-    return getApiForRTDB(context)
-  }
-
-  return getApiForCF(context)
-}
-
-flamelink._registerModule('nav', navigation)
+import '@flamelink/sdk-navigation-cf'
+import '@flamelink/sdk-navigation-rtdb'
