+++ packages/flamelink/rtdb/navigation/index.ts
@@ -1,2 +1,1 @@
-import '@flamelink/sdk-navigation-cf'
 import '@flamelink/sdk-navigation-rtdb'
