+++ packages/flamelink/rtdb/navigation/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "flamelink/rtdb/navigation",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "typings": "../../public.d.ts"
+}
