+++ packages/navigation-cf/README.md
@@ -1,0 +1,5 @@
+# @flamelink/sdk-navigation-cf
+
+This is the Cloud Firestore navigation module for the Flamelink JavaScript SDK.
+
+> This package should only be used via the official [flamelink](https://www.npmjs.com/package/flamelink) package.
