+++ packages/navigation-cf/src/constants.ts
@@ -1,0 +1,7 @@
+export const BATCH_WRITE_LIMIT = 500
+
+export const REQUIRED_FIELDS_FOR_STRUCTURING = [
+  'uuid',
+  'parentIndex',
+  'children'
+]
