+++ packages/navigation-cf/src/vendor.d.ts
