+++ packages/navigation-rtdb/jest.config.js
@@ -1,0 +1,3 @@
+const getJestConfig = require('../../tools/jest-package-base')
+
+module.exports = getJestConfig()
