+++ packages/navigation-rtdb/src/vendor.d.ts
@@ -1,0 +1,1 @@
+declare module 'compose-then'
