--- packages/navigation/src/constants.ts
@@ -1,1 +1,0 @@
-export const CF_BATCH_WRITE_LIMIT = 500
